refactor: switch migration from module to package
