+++ cognite_toolkit/_cdf_tk/commands/_migrate/__init__.py
@@ -1,0 +1,3 @@
+from .timeseries import MigrateTimeseriesCommand
+
+__all__ = ["MigrateTimeseriesCommand"]
