+++ cognite_toolkit/_cdf_tk/commands/_migrate/timeseries.py
@@ -1,8 +1,6 @@
 import csv
-from collections.abc import Collection, Iterator, Sequence
-from dataclasses import dataclass
+from collections.abc import Iterator
 from pathlib import Path
-from typing import SupportsIndex, overload
 
 import questionary
 from cognite.client.data_classes.capabilities import (
@@ -22,7 +20,7 @@ from rich.progress import track
 
 from cognite_toolkit._cdf_tk.client import ToolkitClient
 from cognite_toolkit._cdf_tk.client.data_classes.extended_timeseries import ExtendedTimeSeries
-from cognite_toolkit._cdf_tk.client.data_classes.pending_instances_ids import PendingInstanceId
+from cognite_toolkit._cdf_tk.commands._base import ToolkitCommand
 from cognite_toolkit._cdf_tk.constants import DMS_INSTANCE_LIMIT_MARGIN
 from cognite_toolkit._cdf_tk.exceptions import (
     AuthenticationError,
@@ -33,79 +31,7 @@ from cognite_toolkit._cdf_tk.tk_warnings import HighSeverityWarning
 from cognite_toolkit._cdf_tk.utils import humanize_collection
 from cognite_toolkit._cdf_tk.utils.collection import chunker_sequence
 
-from ._base import ToolkitCommand
-
-
-@dataclass
-class MigrationMapping:
-    resource_type: str
-    instance_id: NodeId
-
-
-@dataclass
-class IdMigrationMapping(MigrationMapping):
-    id: int
-    data_set_id: int | None = None
-
-
-@dataclass
-class ExternalIdMigrationMapping(MigrationMapping):
-    external_id: str
-    data_set_id: int | None = None
-
-
-class MigrationMappingList(list, Sequence[MigrationMapping]):
-    # Implemented to get correct type hints
-    def __init__(self, collection: Collection[MigrationMapping] | None = None) -> None:
-        super().__init__(collection or [])
-
-    def __iter__(self) -> Iterator[MigrationMapping]:
-        return super().__iter__()
-
-    @overload
-    def __getitem__(self, index: SupportsIndex) -> MigrationMapping: ...
-
-    @overload
-    def __getitem__(self, index: slice) -> "MigrationMappingList": ...
-
-    def __getitem__(self, index: SupportsIndex | slice, /) -> "MigrationMapping | MigrationMappingList":
-        if isinstance(index, slice):
-            return MigrationMappingList(super().__getitem__(index))
-        return super().__getitem__(index)
-
-    def get_ids(self) -> list[int]:
-        """Return a list of IDs from the migration mappings."""
-        return [mapping.id for mapping in self if isinstance(mapping, IdMigrationMapping)]
-
-    def get_external_ids(self) -> list[str]:
-        """Return a list of external IDs from the migration mappings."""
-        return [mapping.external_id for mapping in self if isinstance(mapping, ExternalIdMigrationMapping)]
-
-    def as_node_ids(self) -> list[NodeId]:
-        """Return a list of NodeIds from the migration mappings."""
-        return [mapping.instance_id for mapping in self]
-
-    def spaces(self) -> set[str]:
-        """Return a set of spaces from the migration mappings."""
-        return {mapping.instance_id.space for mapping in self}
-
-    def as_pending_ids(self) -> list[PendingInstanceId]:
-        return [
-            PendingInstanceId(
-                pending_instance_id=mapping.instance_id,
-                id=mapping.id if isinstance(mapping, IdMigrationMapping) else None,
-                external_id=mapping.external_id if isinstance(mapping, ExternalIdMigrationMapping) else None,
-            )
-            for mapping in self
-        ]
-
-    def get_data_set_ids(self) -> set[int]:
-        """Return a list of data set IDs from the migration mappings."""
-        return {
-            mapping.data_set_id
-            for mapping in self
-            if isinstance(mapping, IdMigrationMapping | ExternalIdMigrationMapping) and mapping.data_set_id is not None
-        }
+from .data_classes import ExternalIdMigrationMapping, IdMigrationMapping, MigrationMappingList
 
 
 class MigrateTimeseriesCommand(ToolkitCommand):
